requirements.txt
addict
future
lmdb
numpy
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
torch==1.10.0
torchvision
tqdm
yapf
h5py
einops